Move plan upsell to separate component
