+++ client/my-sites/domains/domain-management/edit/domain-types/registered-domain-type.jsx
@@ -23,14 +23,12 @@ import {
 } from 'my-sites/domains/paths';
 import IcannVerificationCard from 'my-sites/domains/domain-management/components/icann-verification';
 import { isRecentlyRegistered, isExpiringSoon } from 'lib/domains/utils';
-import { DOMAINS, SETTING_PRIMARY_DOMAIN } from 'lib/url/support';
+import { DOMAINS } from 'lib/url/support';
 import SubscriptionSettings from '../card/subscription-settings';
 import { recordPaymentSettingsClick } from '../payment-settings-analytics';
 
 import CompactFormToggle from 'components/forms/form-toggle/compact';
-import { Banner } from 'components/banner';
-import { currentUserHasFlag, getCurrentUser } from 'state/current-user/selectors';
-import { NON_PRIMARY_DOMAINS_TO_FREE_USERS } from 'state/current-user/constants';
+import NonPrimaryDomainPlanUpsell from '../../components/domain/non-primary-domain-plan-upsell';
 
 class RegisteredDomainType extends React.Component {
 	getVerticalNavigation() {
@@ -256,42 +254,13 @@ class RegisteredDomainType extends React.Component {
 	}
 
 	planUpsellForNonPrimaryDomain() {
-		const { domain, hasNonPrimaryDomainsFlag, isDomainOnly, selectedSite, translate } = this.props;
-
-		if (
-			! hasNonPrimaryDomainsFlag ||
-			isDomainOnly ||
-			! domain.pointsToWpcom ||
-			domain.isPrimary ||
-			domain.isWPCOMDomain ||
-			domain.isWpcomStagingDomain
-		) {
-			return null;
-		}
+		const { domain, isDomainOnly, selectedSite } = this.props;
 
 		return (
-			<Banner
-				title={ translate( 'This domain is being forwarded to %(primaryDomain)s', {
-					args: {
-						primaryDomain: selectedSite.slug,
-					},
-				} ) }
-				description={ translate(
-					'Upgrade to a paid plan to make {{strong}}%(domain)s{{/strong}} the primary address that your visitors see when they visit your site. {{a}}Learn more{{/a}}',
-					{
-						args: {
-							domain: domain.name,
-						},
-						components: {
-							a: <a href={ SETTING_PRIMARY_DOMAIN } target="_blank" rel="noopener noreferrer" />,
-							strong: <strong />,
-						},
-					}
-				) }
-				callToAction={ translate( 'Upgrade' ) }
+			<NonPrimaryDomainPlanUpsell
 				tracksImpressionName="calypso_non_primary_domain_settings_plan_upsell_impression"
 				tracksClickName="calypso_non_primary_domain_settings_plan_upsell_click"
-				event="calypso_non_primary_domain_settings_plan_upsell"
+				{ ...{ domain, isDomainOnly, selectedSite } }
 			/>
 		);
 	}
@@ -362,15 +331,7 @@ class RegisteredDomainType extends React.Component {
 	}
 }
 
-const mapStateToProps = state => {
-	return {
-		hasNonPrimaryDomainsFlag: getCurrentUser( state )
-			? currentUserHasFlag( state, NON_PRIMARY_DOMAINS_TO_FREE_USERS )
-			: false,
-	};
-};
-
-export default connect( mapStateToProps, {
+export default connect( null, {
 	recordTracksEvent,
 	recordGoogleEvent,
 	recordPaymentSettingsClick,
